+++ lib/UnoCore/UnoCore.unoproj
@@ -48,27 +48,27 @@
     "android/run.bat:File",
     "android/run.sh:File",
     "android/settings.gradle:File",
-    "android/Uno/Base/Bootstrapper.cpp.uxl:Extensions",
-    "android/Uno/Base/Bootstrapper.uno:Source",
-    "android/Uno/Base/BootstrapperImpl_Android.cpp:File",
-    "android/Uno/Base/BootstrapperImpl_Android.h:File",
-    "android/Uno/Base/DirectBuffer.uno:Source",
-    "android/Uno/Base/JNI.cpp.uxl:Extensions",
-    "android/Uno/Base/JNI.uno:Source",
-    "android/Uno/Base/Primitives.cpp.uxl:Extensions",
-    "android/Uno/Base/Primitives.uno:Source",
-    "android/Uno/Base/Resources.uno:Source",
-    "android/Uno/Base/Types.cpp.uxl:Extensions",
-    "android/Uno/Base/Types.uno:Source",
-    "android/Uno/Base/UnoHelper.java:File",
-    "android/Uno/Base/Versions.uno:Source",
-    "android/Uno/Base/Wrappers.cpp.uxl:Extensions",
-    "android/Uno/Base/Wrappers.uno:Source",
-    "android/Uno/EntryPoints.cpp:File",
-    "android/Uno/Graphics/GLHelper.cpp:File",
-    "android/Uno/Graphics/GLHelper.h:File",
-    "android/Uno/JNIHelper.cpp:File",
-    "android/Uno/JNIHelper.h:File",
+    "android/uDroid/Bootstrapper.cpp.uxl:Extensions",
+    "android/uDroid/Bootstrapper.uno:Source",
+    "android/uDroid/BootstrapperImpl_Android.cpp:File",
+    "android/uDroid/BootstrapperImpl_Android.h:File",
+    "android/uDroid/DirectBuffer.uno:Source",
+    "android/uDroid/EntryPoints.cpp:File",
+    "android/uDroid/GLHelper.cpp:File",
+    "android/uDroid/GLHelper.h:File",
+    "android/uDroid/JNI.cpp.uxl:Extensions",
+    "android/uDroid/JNI.uno:Source",
+    "android/uDroid/JNIHelper.cpp:File",
+    "android/uDroid/JNIHelper.h:File",
+    "android/uDroid/Primitives.cpp.uxl:Extensions",
+    "android/uDroid/Primitives.uno:Source",
+    "android/uDroid/Resources.uno:Source",
+    "android/uDroid/Types.cpp.uxl:Extensions",
+    "android/uDroid/Types.uno:Source",
+    "android/uDroid/UnoHelper.java:File",
+    "android/uDroid/Versions.uno:Source",
+    "android/uDroid/Wrappers.cpp.uxl:Extensions",
+    "android/uDroid/Wrappers.uno:Source",
     "assets/adaptive_icon/drawable/icon_background.xml:File",
     "assets/adaptive_icon/drawable/icon_foreground.xml:File",
     "assets/adaptive_icon/mipmap-anydpi-v26/icon_round.xml:File",
