+++ lib/UnoCore/android/uDroid/BootstrapperImpl_Android.cpp
