+++ lib/UnoCore/android/uDroid/EntryPoints.cpp
@@ -3,8 +3,8 @@
 #include <uBase/BaseLib.h>
 #include <XliPlatform/PlatformSpecific/Android.h>
 
-#include <Uno/JNIHelper.h>
-#include <Uno/Graphics/GLHelper.h>
+#include <uDroid/GLHelper.h>
+#include <uDroid/JNIHelper.h>
 
 @{Android.Base.JNI:IncludeDirective}
 @{Uno.Platform.CoreApp:IncludeDirective}
