+++ lib/UnoCore/android/uDroid/JNIHelper.h
