+++ lib/UnoCore/android/uDroid/Resources.uno
