+++ lib/UnoCore/android/uDroid/UnoHelper.java
